requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.1
cssselect==1.1.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
entrypoints==0.4
exceptiongroup==1.0.0rc9
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.2.2
Flask-Classful==0.14.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.16.1
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
mysqlclient==2.1.1
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.23.4
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.15.0
prompt-toolkit==3.0.31
Protego==0.2.1
psutil==5.9.3
psycopg2==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.0
selenium==4.5.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
terminado==0.16.0
tinycss2==1.2.1
tldextract==3.4.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.8.0
typing_extensions==4.4.0
tzdata==2022.4
urllib3==1.26.12
w3lib==2.0.1
wcwidth==0.2.5
webdriver-manager==3.8.4
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
whitenoise==6.2.0
widgetsnbextension==4.0.3
wsproto==1.2.0
zope.interface==5.5.0
